pyproject.toml
[project]
name = "tire-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "coacd>=1.0.6",
    "ezbullet>=0.1.1",
    "gymnasium>=1.1.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.9.4",
    "numpy<2",
    "pyqt5>=5.15.11",
    "shapely>=2.0.7",
]

[tool.setuptools]
packages = ["tire_env"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
